restructure file architecture
